refactor: data 모듈 mapper 네이밍 변경, mapper 패키지 분리
