--- data/src/main/java/com/paykidscompose/data/mapper/TestMapper.kt
@@ -1,4 +1,0 @@
-package com.paykidscompose.data.mapper
-
-class TestMapper {
-}
