+++ data/src/main/java/com/paykidscompose/data/mapper/quiz/QuizClearedMapper.kt
@@ -1,10 +1,10 @@
-package com.paykidscompose.data.mapper
+package com.paykidscompose.data.mapper.quiz
 
 import com.paykidscompose.common.mapper.ModelMapper
 import com.paykidscompose.common.model.QuizClearedModel
 import com.paykidscompose.data.model.quiz.QuizClearedDTO
 
-object QuizClearedDTOMapper : ModelMapper<QuizClearedModel, QuizClearedDTO> {
+object QuizClearedMapper : ModelMapper<QuizClearedModel, QuizClearedDTO> {
     override fun mapToModel(layerModel: QuizClearedDTO): QuizClearedModel {
         return QuizClearedModel(
             message = layerModel.message,
