+++ data/src/main/java/com/paykidscompose/data/mapper/study/ChatResponseMapper.kt
@@ -1,10 +1,10 @@
-package com.paykidscompose.data.mapper
+package com.paykidscompose.data.mapper.study
 
 import com.paykidscompose.common.mapper.ModelMapper
 import com.paykidscompose.common.model.ChatResponseModel
 import com.paykidscompose.data.model.study.ChatResponseDTO
 
-object ChatResponseDTOMapper : ModelMapper<ChatResponseModel, ChatResponseDTO> {
+object ChatResponseMapper : ModelMapper<ChatResponseModel, ChatResponseDTO> {
     override fun mapToModel(layerModel: ChatResponseDTO): ChatResponseModel {
         return ChatResponseModel(
             response = layerModel.response
